requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pydantic==2.7.1
google-api-python-client==2.130.0
google-auth==2.30.0
psycopg[binary,pool]==3.1.18
python-dotenv==1.0.1
openai==1.1.0
